[BACKEND-8] Address PR review comments
