+++ src/modules/database/schema/university-admin.table.ts
@@ -4,11 +4,8 @@ export const universityAdmin = pgTable(
   'universityadmin',
   {
     adminId: uuid('admin_id').primaryKey().defaultRandom(),
-
     userId: uuid('user_id').notNull(),
-
     universityId: uuid('university_id').notNull(),
-
     createdAt: timestamp('created_at', { withTimezone: false }).notNull().defaultNow(),
     updatedAt: timestamp('updated_at', { withTimezone: false }).notNull().defaultNow(),
   },
