--- src/modules/university-admin/index.ts
@@ -1,2 +1,0 @@
-export * from './university-admin.module';
-export * from './university-admin.repository';
