+++ src/modules/university_admin/index.ts
@@ -1,0 +1,2 @@
+export { UniversityAdminModule } from './university_admin.module';
+export { UniversityAdminRepository } from './university_admin.repository';
