+++ src/modules/university_admin/university_admin.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from '../database/database.module';
-import { UniversityAdminRepository } from './university-admin.repository';
+import { UniversityAdminRepository } from './university_admin.repository';
 
 @Module({
   imports: [DatabaseModule],
