+++ src/shared/types/repository/university_admin.repository.types.ts
@@ -1,0 +1,8 @@
+export type CreateUniversityAdminInput = {
+  userId: string;
+  universityId: string;
+};
+
+export type UpdateUniversityAdminInput = {
+  universityId?: string;
+};
